JSON format GUI
This file is the JSON format for GUI which is sent to the back end. 
To use import this file in your node editor, follow the instruction:
1. Copy the content of the file,
2. Start the node-red server and import  the copied clipboard content using ctrl + I and paste, import
The flow must appear automatically with three  nodes "CSharplib", "JSON2doublematrix" and "Csharpweibullalg".
Currently there is no function embedded into these three nodes.